pyproject.toml
[project]
name = "1stllm"
version = "0.1.0"
description = "Gemini AI Tool-Calling Chatbot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "python-dotenv",
    "prompt_toolkit",
    "requests>=2.32.5",
    "rdkit>=2025.9.6",
]

# Move your dev tools here
[dependency-groups]
dev = [
    "pytest",
]
